Fix for u32 and f64 type mismatches in arcard-1-8

+++ code/arcaders-1-8/src/phi/mod.rs
@@ -38,7 +38,7 @@ impl<'window> Phi<'window> {
         }
     }
 
-    pub fn output_size(&self) -> (u32, u32) {
+    pub fn output_size(&self) -> (f64, f64) {
         let (w, h) = self.renderer.output_size().unwrap();
         (w as f64, h as f64)
     }
